+++ src/main/java/io/hhplus/tdd/common/ErrorResponse.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.common;
 
 public record ErrorResponse(
         String code,
