+++ src/main/java/io/hhplus/tdd/point/model/PointHistory.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.model;
 
 public record PointHistory(
         long id,
