+++ src/main/java/io/hhplus/tdd/point/model/UserPoint.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.model;
 
 public record UserPoint(
         long id,
